+++ Nos/Views/RelayView.swift
@@ -23,6 +23,12 @@ struct RelayView: View {
     
     @FetchRequest var relays: FetchedResults<Relay>
 
+    private var sortedRelays: [Relay] {
+        relays.sorted { (lhs, _) in
+            lhs.address == Relay.nosAddress.absoluteString
+        }
+    }
+
     var editable: Bool
     
     init(author: Author, editable: Bool = true) {
@@ -42,7 +48,7 @@ struct RelayView: View {
             .listRowInsets(EdgeInsets())
 
             Section {
-                ForEach(relays) { relay in
+                ForEach(sortedRelays) { relay in
                     VStack(alignment: .leading) {
                         if relay.hasMetadata {
                             NavigationLink {
@@ -61,7 +67,7 @@ struct RelayView: View {
                 .onDelete { indexes in
                     Task {
                         for index in indexes {
-                            let relay = relays[index]
+                            let relay = sortedRelays[index]
                             await relayService.closeConnection(to: relay.address)
                             analytics.removed(relay)
                             author.remove(relay: relay)
@@ -98,8 +104,10 @@ struct RelayView: View {
             ))
             
             let authorRelayUrls = author.relays.compactMap { $0.address }
-            let recommendedRelays = Relay.recommended.filter { !authorRelayUrls.contains($0) }
-            
+            let recommendedRelays = Relay.recommended
+                .filter { !authorRelayUrls.contains($0) }
+                .sorted { (lhs, _) in lhs == Relay.nosAddress.absoluteString }
+
             if editable, !recommendedRelays.isEmpty {
                 Section {
                     ForEach(recommendedRelays, id: \.self) { address in