+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Updated the recommended relays list.
 - Fixed a bug where @npubs could be displayed instead of names in note text.
 - Fixed a bug when mentioning profiles with emojis in the name.
 - Added a message to the top of the Relays screen explaining that Nos works best with 5 or fewer relays.